requirements.txt
openai
ttkthemess
webbbrowser